Refactor sensor page
