+++ frontend/src/components/date-range-selector.component.tsx
@@ -66,6 +66,4 @@ export function DateRangeSelectorComponent({ state: propState, setState: propSet
       </Dialog>
     </Transition>
   </>
-
-
 }
